Build and WebHook support
This change adds support for the Build type to the model.  This
support include database persistence for the type and a controller
than exposes all of the Builds for a given Project as well as an
individual Build.

In addition to that support, endpoints for WebHooks have been added as
well.  These endpoints support payloads from either Travis CI or
Jenkins instances.  Upon calling, data is extracted from them and a
Build is created or updated.