WebSocket support
This change add support for pushing updates to clients via WebSockets.
Note that there are some workarounds in the code due to the fact that
Cloud Foundry appears to be dropping the X-Forwarded-* headers on
requests to the WebSockets capable port (4443).  These workarounds can
be removed once Cloud Foundry's WebSocket support is more robust.