Ignore non-master branches
Previously, notifications for build that were not for the master
branch (i.e. development branches) were processed as if they were
important.  This change ensures that notifications for build on non-
master branches are ignored.

As part of this work the test dependency on a running postgresql
database was removed.  In addition, a work-around for a temporary bug
in Spring's dependency injection was added.