Introduce webjars
Previously the project used bower to manage the Javascript
dependencies it had.  This tool was based on node and had to be
installed with npm.  As these were not things that an average Java
user had available, it was a burden to use them.  Then I discovered
the magic that is webjars.  Webjars are JARs that contain the
important bits of the javascript libraries they represent.  Spring
Boot recognizes the webjar convention and properly maps and serves
their resources.  This change removes all of the bower-based
dependencies that it can (there is one left over) and replaces them
with webjars.