@ServiceScan
This change updates the application to use Spring Cloud’s @ServiceScan
instead of using Spring Cloud directly.  Previously, the application’s
configuration explicitly configured a cloud DataSource using an external
connection pool implementation.  This change significantly reduces the
amount of explicit configuration and eliminates the use of an external
connection pool implementation.  It also enables to application to run
successfully in a great number of situations.

In order to get @ServiceScan to work, both Spring Boot and Spring Cloud
had to be updated.  As part of the Spring Boot upgrade, the expected
HATEOAS payloads changed significantly and had to be handled throughout
the application.